ifconfig: T2104: splt ifconfig.py into multiple files
